chore: replaced unification with implicit usage
